feat: update ssr-vite skip ci
