+++ .npmrc
@@ -11,10 +11,10 @@ public-hoist-pattern[]=*vue-server-renderer*
 public-hoist-pattern[]=ssr-serialize-javascript
 public-hoist-pattern[]=*@vue/server-renderer*
 public-hoist-pattern[]=*ssr*
+public-hoist-pattern[]=ssr-vite
 public-hoist-pattern[]=pinia
 public-hoist-pattern[]=@types/webpack
 public-hoist-pattern[]=@types/react
 public-hoist-pattern[]=axios
 public-hoist-pattern[]=*core-js*
 public-hoist-pattern[]=path-to-regexp
-public-hoist-pattern[]=*ssr*
