+++ example/midway-vue3-ssr/package.json
@@ -28,6 +28,8 @@
     "ssr-plugin-midway": "workspace:*",
     "ssr-plugin-vue3": "workspace:*",
     "ssr-types": "workspace:*",
+    "ssr-vite": "workspace:*",
+    "ssr-webpack": "workspace:*",
     "typescript": "^5.0.0"
   },
   "scripts": {
