+++ packages/core/package.json
@@ -31,9 +31,10 @@
     "cheerio": "^1.0.0-rc.10",
     "ssr-serialize-javascript": "^6.0.0",
     "ssr-deepclone": "^1.0.0",
-    "ssr-common-utils": "^6.2.10"
+    "ssr-common-utils": "workspace:*"
   },
   "devDependencies": {
-    "ssr-types": "^6.0.0"
+    "ssr-types": "workspace:*",
+    "ssr-vite": "workspace:*"
   }
 }
