+++ packages/plugin-vue3/package.json
@@ -31,12 +31,10 @@
   "homepage": "https://github.com/zhangyuang/ssr#readme",
   "dependencies": {
     "ssr-common-utils": "workspace:*",
-    "rollup-plugin-visualizer": "^5.12.0",
     "@vitejs/plugin-vue": "^2.3.4",
     "@vitejs/plugin-vue-jsx": "^1.3.2",
     "@vue/babel-plugin-jsx": "^1.0.3",
     "@vue/compiler-sfc": "^3.0.7",
-    "vite": "^2.9.15",
     "vue": "^3.0.0",
     "vue-loader": "^17.0.0",
     "vue-router": "^4.2.4",
