+++ packages/plugin-vue3/src/index.ts
@@ -1,51 +1,47 @@
 import { resolve } from 'path'
 import { loadConfig, getCwd } from 'ssr-common-utils'
-import type { webpackStart as webpackStartType } from 'ssr-webpack'
 
 const { isVite, optimize } = loadConfig()
 const spinner = require('ora')('Building')
 
-export function clientPlugin() {	
+export function clientPlugin() {
 	const cwd = getCwd()
 	const webpackPath = resolve(cwd, './node_modules/ssr-webpack')
+	const vitePath = resolve(cwd, './node_modules/ssr-vite')
 	return {
 		name: 'plugin-vue3',
 		start: async () => {
 			if (isVite) {
-				const { viteStart } = await import('./tools/vite')
-				await viteStart()
+				const { start } = await import(vitePath)
+				await start()
 			} else {
 				if (optimize) {
 					spinner.start()
-					const { viteBuildClient } = await import('./tools/vite')
+					const { viteBuildClient } = await import(vitePath)
 					await viteBuildClient()
 					process.env.NODE_ENV = 'development'
 					spinner.stop()
 				}
-				const { webpackStart }: { webpackStart: typeof webpackStartType } = await import(webpackPath) 
-				await webpackStart()
+				const { start } = await import(webpackPath)
+				await start()
 			}
 		},
 		build: async () => {
 			if (isVite) {
-				const { viteBuild } = await import('./tools/vite')
-				await viteBuild()
+				const { build } = await import(vitePath)
+				await build()
 			} else {
 				if (optimize) {
 					spinner.start()
-					const { viteBuildClient } = await import('./tools/vite')
-					await viteBuildClient()
+					const { build } = await import(vitePath)
+					await build()
 					spinner.stop()
-					const { webpackBuild } = await import(webpackPath)
-					await webpackBuild()
-				} else {
-					const { webpackBuild } = await import(webpackPath)
-					await webpackBuild()
 				}
+				const { build } = await import(webpackPath)
+				await build()
 			}
 		}
 	}
 }
 
-export * from './tools/vite'
 export * from './types'
