+++ packages/vite/package.json
@@ -27,7 +27,9 @@
   },
   "homepage": "https://github.com/zhangyuang/ssr#readme",
   "dependencies": {
-  
+    "vite": "^2.9.15",
+    "rollup-plugin-visualizer": "^5.12.0",
+    "ssr-vite-plugin-style-import": "^2.0.1"
   },
   "devDependencies": {
     "ssr-types": "workspace:*"
