+++ packages/webpack/src/index.ts
@@ -2,7 +2,7 @@ import * as WebpackChain from 'webpack-chain'
 import { startClientServer, startClientBuild } from './client'
 import { startServerBuild } from './server'
 
-export const webpackStart = async () => {
+export const start = async () => {
 	const serverConfigChain = new WebpackChain()
 	const clientConfigChain = new WebpackChain()
 
@@ -12,7 +12,7 @@ export const webpackStart = async () => {
 	await Promise.all([startServerBuild(getServerWebpack(serverConfigChain)), startClientServer(getClientWebpack(clientConfigChain))])
 }
 
-export const webpackBuild = async () => {
+export const build = async () => {
 	const { getServerWebpack } = await import('./config/server-config')
 	const { getClientWebpack } = await import('./config/client-config')
 	const serverConfigChain = new WebpackChain()
